Removed  other persistence adapter and left only "sleep" one, to reduce the size of the distribution artifact, allowing for a smoother tutorial/demo experience.

+++ Components/DaaSWebClient/gangliaPlugIns/getThroughput.py
@@ -36,7 +36,7 @@ def metric_init(params):
     d1 = {'name': 'throughput',
         'call_back': temp_handler,
         'time_max': 5,
-        'value_type': 'double',
+        'value_type': 'int',
  'units': '#',
         'slope': 'both',
         'format': '%d',
